[DataObjects] Refactor multiple assignments check for Advanced relation fix - review changes #5902
