+++ models/DataObject/ClassDefinition/Data/AdvancedManyToManyObjectRelation.php
@@ -603,7 +603,7 @@ class AdvancedManyToManyObjectRelation extends ManyToManyObjectRelation
 
         $objectsMetadata = $this->getDataFromObjectParam($object, $params);
         //TODO: move validation to checkValidity & throw exception in Pimcore 7
-        if ($object->isFieldDirty($this->getName()) || $object->isFieldDirty('_self')) {
+        if (!$object instanceof DataObject\DirtyIndicatorInterface || $object->isFieldDirty($this->getName()) || $object->isFieldDirty('_self')) {
             $objectsMetadata = $this->filterMultipleAssignments($objectsMetadata, $object, $params);
         }
