+++ models/DataObject/ClassDefinition/Data/ManyToManyObjectRelation.php
@@ -26,6 +26,7 @@ class ManyToManyObjectRelation extends AbstractRelations implements QueryResourc
     use Model\DataObject\ClassDefinition\Data\Extension\Relation;
     use Extension\QueryColumnType;
     use DataObject\ClassDefinition\Data\Relations\AllowObjectRelationTrait;
+    use DataObject\ClassDefinition\Data\Relations\ManyToManyRelationTrait;
 
     /**
      * Static type of this element
@@ -518,37 +519,6 @@ class ManyToManyObjectRelation extends AbstractRelations implements QueryResourc
         }
 
         return $relatedObjects;
-    }
-
-    /**
-     * TODO: move validation to checkValidity & throw exception in Pimcore 7
-     * @param DataObject\Concrete|DataObject\Localizedfield|DataObject\Objectbrick\Data\AbstractData|\Pimcore\Model\DataObject\Fieldcollection\Data\AbstractData $object
-     * @param array $params
-     */
-    public function save($object, $params = [])
-    {
-        if (!DataObject\AbstractObject::isDirtyDetectionDisabled() && $object instanceof DataObject\DirtyIndicatorInterface) {
-            if ($object instanceof DataObject\Localizedfield) {
-                if ($object->getObject() instanceof DataObject\DirtyIndicatorInterface) {
-                    if (!$object->hasDirtyFields()) {
-                        return;
-                    }
-                }
-            } else {
-                if ($this->supportsDirtyDetection()) {
-                    if (!$object->isFieldDirty($this->getName())) {
-                        return;
-                    }
-                }
-            }
-        }
-
-        $objectsMetadata = $this->getDataFromObjectParam($object, $params);
-        if ($object->isFieldDirty($this->getName()) || $object->isFieldDirty('_self')) {
-            $this->filterMultipleAssignments($objectsMetadata, $object, $params);
-        }
-
-        parent::save($object, $params);
     }
 
     /**
