+++ models/DataObject/ClassDefinition/Data/ManyToManyRelation.php
@@ -30,6 +30,7 @@ class ManyToManyRelation extends AbstractRelations implements QueryResourcePersi
     use DataObject\ClassDefinition\Data\Relations\AllowObjectRelationTrait;
     use DataObject\ClassDefinition\Data\Relations\AllowAssetRelationTrait;
     use DataObject\ClassDefinition\Data\Relations\AllowDocumentRelationTrait;
+    use DataObject\ClassDefinition\Data\Relations\ManyToManyRelationTrait;
 
     /**
      * Static type of this element
@@ -697,38 +698,6 @@ class ManyToManyRelation extends AbstractRelations implements QueryResourcePersi
         } else {
             throw new \Exception('cannot get values from web service import - invalid data');
         }
-    }
-
-    /**
-     * TODO: move validation to checkValidity & throw exception in Pimcore 7
-     * @param DataObject\Concrete|DataObject\Localizedfield|DataObject\Objectbrick\Data\AbstractData|\Pimcore\Model\DataObject\Fieldcollection\Data\AbstractData $object
-     * @param array $params
-     */
-    public function save($object, $params = [])
-    {
-        if (!DataObject\AbstractObject::isDirtyDetectionDisabled() && $object instanceof DataObject\DirtyIndicatorInterface) {
-            if ($object instanceof DataObject\Localizedfield) {
-                if ($object->getObject() instanceof DataObject\DirtyIndicatorInterface) {
-                    if (!$object->hasDirtyFields()) {
-                        return;
-                    }
-                }
-            } else {
-                if ($this->supportsDirtyDetection()) {
-                    if (!$object->isFieldDirty($this->getName())) {
-                        return;
-                    }
-                }
-            }
-        }
-
-        $multihrefMetadata = $this->getDataFromObjectParam($object, $params);
-
-        if ($object->isFieldDirty($this->getName()) || $object->isFieldDirty('_self')) {
-            $this->filterMultipleAssignments($multihrefMetadata, $object, $params);
-        }
-
-        parent::save($object, $params);
     }
 
     /**
